Ch. 17§02 initial edits on 'Message Passing'
- The usual kinds of clarifications, cleaning up redundancies, etc.
- Fold together listings:
    - 8 and 9 → 8
    - 10, 11, 12 → 9
    - 13 → 10
    - 15 → 11
    - 16 → 12
- Tweak the durations used for the listings throughout this section,
  including making the durations *differ* between the two sending blocks
  in the final example.
