+++ listings/ch17-async-await/listing-17-05/src/main.rs
@@ -5,13 +5,13 @@ fn main() {
         trpl::spawn_task(async {
             for i in 1..10 {
                 println!("hi number {i} from the first task!");
-                trpl::sleep(Duration::from_millis(1)).await;
+                trpl::sleep(Duration::from_millis(500)).await;
             }
         });
 
         for i in 1..5 {
             println!("hi number {i} from the second task!");
-            trpl::sleep(Duration::from_millis(1)).await;
+            trpl::sleep(Duration::from_millis(500)).await;
         }
     });
 }
