+++ listings/ch17-async-await/listing-17-06/src/main.rs
@@ -6,13 +6,13 @@ fn main() {
         let handle = trpl::spawn_task(async {
             for i in 1..10 {
                 println!("hi number {i} from the first task!");
-                trpl::sleep(Duration::from_millis(1)).await;
+                trpl::sleep(Duration::from_millis(500)).await;
             }
         });
 
         for i in 1..5 {
             println!("hi number {i} from the second task!");
-            trpl::sleep(Duration::from_millis(1)).await;
+            trpl::sleep(Duration::from_millis(500)).await;
         }
 
         handle.await.unwrap();
