+++ listings/ch17-async-await/listing-17-07/src/main.rs
@@ -6,14 +6,14 @@ fn main() {
         let fut1 = async {
             for i in 1..10 {
                 println!("hi number {i} from the first task!");
-                trpl::sleep(Duration::from_millis(1)).await;
+                trpl::sleep(Duration::from_millis(500)).await;
             }
         };
 
         let fut2 = async {
             for i in 1..5 {
                 println!("hi number {i} from the second task!");
-                trpl::sleep(Duration::from_millis(1)).await;
+                trpl::sleep(Duration::from_millis(500)).await;
             }
         };
