+++ listings/ch17-async-await/listing-17-08/src/main.rs
@@ -1,13 +1,13 @@
 fn main() {
     trpl::block_on(async {
-        // ANCHOR: add-channel
+        // ANCHOR: channel
         let (tx, mut rx) = trpl::channel();
-        // ANCHOR_END: add-channel
 
         let val = String::from("hi");
         tx.send(val).unwrap();
 
         let received = rx.recv().await.unwrap();
         println!("Got: {received}");
+        // ANCHOR_END: channel
     });
 }
