+++ listings/ch17-async-await/listing-17-10/src/main.rs
@@ -1,26 +1,33 @@
-// ANCHOR: many-messages
 use std::time::Duration;
-// ANCHOR_END: many-messages
 
 fn main() {
     trpl::block_on(async {
         let (tx, mut rx) = trpl::channel();
 
-        // ANCHOR: many-messages
+        // ANCHOR: futures
+        let tx_fut = async {
+            let vals = vec![
+                String::from("hi"),
+                String::from("from"),
+                String::from("the"),
+                String::from("future"),
+            ];
 
-        // snip...
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_millis(500)).await;
+            }
+        };
 
-        let vals = vec![
-            String::from("hi"),
-            String::from("from"),
-            String::from("the"),
-            String::from("future"),
-        ];
+        let rx_fut = async {
+            // ANCHOR: loop
+            while let Some(value) = rx.recv().await {
+                println!("received '{value}'");
+            }
+            // ANCHOR_END: loop
+        };
 
-        for val in vals {
-            tx.send(val).unwrap();
-            trpl::sleep(Duration::from_secs(1)).await;
-        }
-        // ANCHOR_END: many-messages
+        trpl::join(tx_fut, rx_fut).await;
+        // ANCHOR_END: futures
     });
 }
