+++ listings/ch17-async-await/listing-17-11/src/main.rs
@@ -2,24 +2,31 @@ use std::time::Duration;
 
 fn main() {
     trpl::block_on(async {
+        // ANCHOR: with-move
         let (tx, mut rx) = trpl::channel();
 
-        let vals = vec![
-            String::from("hi"),
-            String::from("from"),
-            String::from("the"),
-            String::from("future"),
-        ];
+        let tx_fut = async move {
+            let vals = vec![
+                String::from("hi"),
+                String::from("from"),
+                String::from("the"),
+                String::from("future"),
+            ];
 
-        for val in vals {
-            tx.send(val).unwrap();
-            trpl::sleep(Duration::from_secs(1)).await;
-        }
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_millis(500)).await;
+            }
+        };
 
-        // ANCHOR: loop
-        while let Some(value) = rx.recv().await {
-            println!("received '{value}'");
-        }
-        // ANCHOR_END: loop
+        let rx_fut = async {
+            while let Some(value) = rx.recv().await {
+                eprintln!("received '{value}'");
+            }
+        };
+
+        trpl::join(tx_fut, rx_fut).await;
+        // ANCHOR_END: with-move
     });
 }
+// ANCHOR_END: all
