+++ listings/ch17-async-await/listing-17-12/src/main.rs
@@ -1,25 +1,46 @@
-// ANCHOR: all
 use std::time::Duration;
 
 fn main() {
     trpl::block_on(async {
+        // ANCHOR: here
         let (tx, mut rx) = trpl::channel();
 
-        let vals = vec![
-            String::from("hi"),
-            String::from("from"),
-            String::from("the"),
-            String::from("future"),
-        ];
+        let tx1 = tx.clone();
+        let tx1_fut = async move {
+            let vals = vec![
+                String::from("hi"),
+                String::from("from"),
+                String::from("the"),
+                String::from("future"),
+            ];
 
-        for val in vals {
-            tx.send(val).unwrap();
-            trpl::sleep(Duration::from_secs(1)).await;
-        }
+            for val in vals {
+                tx1.send(val).unwrap();
+                trpl::sleep(Duration::from_millis(500)).await;
+            }
+        };
 
-        while let Some(value) = rx.recv().await {
-            println!("received '{value}'");
-        }
+        let rx_fut = async {
+            while let Some(value) = rx.recv().await {
+                println!("received '{value}'");
+            }
+        };
+
+        let tx_fut = async move {
+            let vals = vec![
+                String::from("more"),
+                String::from("messages"),
+                String::from("for"),
+                String::from("you"),
+            ];
+
+            for val in vals {
+                tx.send(val).unwrap();
+                trpl::sleep(Duration::from_millis(1500)).await;
+            }
+        };
+
+        trpl::join3(tx1_fut, tx_fut, rx_fut).await;
+        // ANCHOR_END: here
     });
 }
-// ANCHOR_END: all
