--- listings/ch17-async-await/listing-17-14/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "async_await"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-trpl = { path = "../../../packages/trpl" }
