--- listings/ch17-async-await/listing-17-15/src/main.rs
@@ -1,32 +1,0 @@
-use std::time::Duration;
-
-fn main() {
-    trpl::block_on(async {
-        // ANCHOR: with-move
-        let (tx, mut rx) = trpl::channel();
-
-        let tx_fut = async move {
-            let vals = vec![
-                String::from("hi"),
-                String::from("from"),
-                String::from("the"),
-                String::from("future"),
-            ];
-
-            for val in vals {
-                tx.send(val).unwrap();
-                trpl::sleep(Duration::from_secs(1)).await;
-            }
-        };
-
-        let rx_fut = async {
-            while let Some(value) = rx.recv().await {
-                eprintln!("received '{value}'");
-            }
-        };
-
-        trpl::join(tx_fut, rx_fut).await;
-        // ANCHOR_END: with-move
-    });
-}
-// ANCHOR_END: all
